Update UI of chatbox
