+++ frontend/src/components/Collaboration/Chat/Chat.js
@@ -1,0 +1,45 @@
+import React, { useState, useEffect } from 'react';
+import './Chat.css';
+
+const Chat = ({ socket, roomId }) => {
+  const [messages, setMessages] = useState([]);
+  const [inputMessage, setInputMessage] = useState('');
+
+  useEffect(() => {
+    socket.on('message', (message) => {
+      setMessages((prevMessages) => [...prevMessages, message]);
+    });
+  }, []);
+
+  const handleInputChange = (e) => {
+    setInputMessage(e.target.value);
+  };
+
+  const handleSendMessage = () => {
+    if (inputMessage) {
+      socket.emit('chatMessage', { room: roomId, message: inputMessage });
+      setInputMessage('');
+    }
+  };
+
+  return (
+    <div className="chat-container">
+      <div className="chat-messages">
+        {messages.map((msg, index) => (
+          <div key={index}>{msg}</div>
+        ))}
+      </div>
+      <div className="chat-input">
+        <input
+          type="text"
+          placeholder="Enter a message"
+          value={inputMessage}
+          onChange={handleInputChange}
+        />
+        <button onClick={handleSendMessage}>Send</button>
+      </div>
+    </div>
+  );
+};
+
+export default Chat;
