+++ frontend/src/pages/Collaboration/Collaboration.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import Chat from '../../components/Collaboration/Chat';
+import Chat from '../../components/Collaboration/Chat/Chat';
 import LiveCode from '../../components/Collaboration/LiveCode';
 import io from 'socket.io-client';
